+++ contracts.py
@@ -4,11 +4,15 @@ from dataclasses import dataclass
 from typing import Any
 
 from fs import FileSystem
+from themer import Themer
+from auth import UserStore
 
 @dataclass
 class CommandContext:
     fs: FileSystem
     username: str
+    themer: Themer
+    user_store: UserStore
 
 @dataclass
 class CommandOutput:
@@ -19,7 +23,3 @@ class CommandOutput:
     @staticmethod
     def text(s: str):
         return CommandOutput(styled=s, plain=s)
-
-    @staticmethod
-    def err(msg: str):
-        return CommandOutput(styled=msg, plain=None, error=True)