+++ shell_exceptions.py
@@ -1,4 +1,0 @@
-# shell_exceptions.py
-
-class ShellExit(Exception):
-    pass